% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAssemble.R
\name{append_qvalues}
\alias{append_qvalues}
\title{Append Adjusted P-Values to Data Frame}
\usage{
append_qvalues(p.combined, correction = "BH")
}
\arguments{
\item{p.combined}{A data frame with combined p-values.}

\item{correction}{A string specifying the p-value correction method.}
}
\description{
Internal helper function to append corrected p-values to the combined data frame.
}
\keyword{internal}
