% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAssemble.R
\name{CCT}
\alias{CCT}
\title{Combine P-Values Using CCT}
\usage{
CCT(pvals, weights = NULL)
}
\arguments{
\item{pvals}{A vector of p-values.}

\item{weights}{Optional weights for the CCT method.}
}
\description{
Internal helper function to combine p-values using the CCT method.
}
\keyword{internal}
