% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dassemble.R
\name{DAssemble}
\alias{DAssemble}
\title{Differential analysis using DAssemble}
\usage{
DAssemble(
  features,
  metadata,
  core_method = NULL,
  enhancers = NULL,
  expVar = "group",
  p_adj = "BY",
  enhancer_norm = "TSS",
  return_components = TRUE,
  return_subensembles = FALSE
)
}
\description{
The \code{DAssemble} function implements an ensemble strategy for differential
analysis. Users can select a single \emph{core} model and optionally augment it
with up to three \emph{enhancer} models. Enhancers are simple, often nonparametric
tests that provide complementary evidence. P-values from the core and
enhancer models are combined using the Cauchy Combination Test (CCT).
}
